+++ frontend/app/assets/styles/auth.css
@@ -76,18 +76,6 @@
 	gap: var(--space-xs);
 }
 
-.auth-modal-backdrop {
-	position: fixed;
-	inset: 0;
-	z-index: 1000;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding: var(--space-l);
-	background-color: var(--overlay-backdrop);
-	backdrop-filter: blur(4px);
-}
-
 .auth-modal-close-row {
 	display: flex;
 	flex-shrink: 0;
