+++ frontend/app/assets/styles/modal.css
@@ -1,0 +1,11 @@
+.modal-backdrop {
+	position: fixed;
+	inset: 0;
+	z-index: 1000;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	padding: var(--space-l);
+	background-color: var(--overlay-backdrop);
+	backdrop-filter: blur(4px);
+}
