+++ frontend/app/locales/ru/translation.json
@@ -65,6 +65,7 @@
 		"title": "Поделитесь впечатлениями об этом рецепте",
 		"rateRecipe": "Оцените этот рецепт",
 		"describeRecipe": "Опишите рецепт (необязательно)",
+		"ratingOption": "{{value}} из 5 звезд",
 		"submit": "Отправить"
 	},
 	"usersPage": {
